// src/app/models/location.models.ts
export interface Location {
    id: number;
    division: string;
    district: string;
    taluk: string;
    village: string;







}
export interface LandAlienation {
    name_of_scheme: string;
    land_alienated: string;
    date_of_enterupon_permission: string;
    sf_number: string;
    go_details: string;
    ownership_land_revenue: string;
    utilization_land: string;
    division: string;
    district: string;
    taluk: string;
    village: string;
  }